refactor(hooks): useFormWorkoutConfigをuseFormConfigにリネームして簡潔に
## 変更内容
- useFormWorkoutConfig.js → useFormConfig.js にリネーム
- 関数名・インポート文をすべて更新
- 状態管理の循環参照問題を解決（前回のコミット内容も含む）

## 理由
- useWorkoutConfigとの区別がしにくかった問題を解決
- フォーム専用の設定フックであることが明確に
- ファイル名がより簡潔で理解しやすく

## 修正内容の詳細
- WorkoutForm.jsx: formConfig参照をworkoutConfigに統一
- ESLint警告の修正（console.log削除、配列keyの修正）
- 不要な間接参照を削除してコードをシンプル化
